Fixes issue 5, 6, 7, 8.
Set clientPortAddress by default, to the ip or hostname the Zookeepers also use to talk to each other.
Add an opt-out property in case people need ZK to bind to all interfaces instead of just that ip/hostname.

Add property for setting dataLogDir. Default it to be the same as the dataDir, and leave the zoo.cfg property unset
unless they are different.

Add property for working around https://issues.apache.org/jira/browse/ZOOKEEPER-3056.
I think it is fine to ask people to deploy ZK twice when doing this upgrade.

Added steps to verify that the ZK cluster looks healthy after deployment. These steps used to be manual.